EditForMe/requirements.txt
google-generativeai==0.4.1
pygame
